+++ server/src/lpcserver/utils.ts
@@ -4,6 +4,7 @@ import * as protocol from "../server/_namespaces/lpc.server.protocol.js";
 import * as typeConverters from './typeConverters.js';
 import * as PConst from "./protocol.const.js";
 import { URI } from "vscode-uri";
+import { IFilePathToResourceConverter } from "./textRendering.js";
 
 const getSymbolKind = (kind: string): ls.SymbolKind => {
 	switch (kind) {
@@ -151,3 +152,7 @@ export function convertNavTree(
     return shouldInclude;
 }
 
+export function getFileResourceConverter(): IFilePathToResourceConverter {    
+    const converter: IFilePathToResourceConverter = { toResource: (file: string) => file && URI.file(file) };
+    return converter;
+}